LA gestion de pacientes, administradores y medicos esta casi lista.

+++ TurboMedical4000-war/web/administradorList.jsp
@@ -46,7 +46,7 @@
                 <td><%= usuario.getIdAdministrador() %></td>
                 <td><%= usuario.getUsuario() %></td>
                 <td><a href="EditAdministradoresServlet?do=editForm&idAdministrador=<%= usuario.getIdAdministrador() %>">Editar</a>
-                    <a href="EditAdministradoresServlet?do=delete&idAdministrador=<%= usuario.getIdAdministrador() %>">Borrar</a></td>
+                    <a href="EditAdministradoresServlet?do=delete&idAdministrador=<%= usuario.getIdAdministrador() %>" onclick="return confirm('¿Seguro que desea eliminar\nel usuario <%= usuario.getUsuario()%>?')">Borrar</a></td>
             </tr>
 
 <%
